+++ caloriechecker/src/Components/AboutUs.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Title from '../Modules/Title';
+// import Css from './homeCss'
+
+function About() {
+  return (
+    <div>
+      <Title />
+    </div>
+  );
+}
+
+export default About;
