+++ caloriechecker/src/Modules/Title.js
@@ -8,13 +8,16 @@ function Title() {
     <div className="title">
       <a href='/' className="link"> <h1 className="titleText">Weightloss Tracker</h1> </a>
       <div className="buttons">
-          
             <button class="button" onClick={() => navigate('/diary')}>
               Diary
             </button>
             |
             <button class="button" onClick={() => navigate('/create')}>
               Create
+            </button>
+            |
+            <button class="button" onClick={() => navigate('/AboutUs')}>
+              AboutUs
             </button>
       </div>
     </div>
