+++ caloriechecker/src/test.api
@@ -1,0 +1,32 @@
+{
+    "food_id": "Mexican Blend Shredded Cheese",
+    "serving_id": "1 ounce",
+    "food_entry_id": "1 ounce of Mexican Blend Shredded Cheese",
+    "dateint": 19402,
+    "food_type": "Brand",
+    "brand_type": "HEB",
+    "meal_type": "",
+    "calories": 110,
+    "carbohydrate": "1",
+    "fiber": "0",
+    "sugar": "0",
+    "added_sugars": "0",
+    "protein": "6",
+    "fat": "8",
+    "saturated_fat": "5",
+    "polyunsaturated_fat": "0",
+    "monounsaturated_fat": "0",
+    "trans_fat": "0",
+    "cholesterol": 25,
+    "sodium": 170,
+    "potassium": 20,
+    "vitamin_d": 0,
+    "vitamin_a": "",
+    "vitamin_c": "",
+    "calcium": 0,
+    "iron": ".1",
+    "targetCalories": 0,
+    "targetProteins": 0,
+    "targetCarbohydrates": 0,
+    "targetFat": 0
+}
